refactor v1 of app.py
